requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
cryptography==42.0.2
deprecation==2.1.0
docker==7.0.0
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0
wrapt==1.16.0
